+++ doc/source/changes.rst
@@ -7,6 +7,8 @@ NEXT
 ****
 * Added interface to allow transporting git data: **TransportDBMixin**
 * Added interface to allow reference resolution: **RefParseMixin**
+* Added interface to handle git related paths: **RepositoryPathsMixin**
+* Added interface to read and write git-like configuration: **ConfigurationMixin**
 * Added implementation of git datbase with support for transportation and reference resolution: **RefGitDB**
 
 *****